# Prevent IndexOutOfRangeException in RegexInterpreter