@stephentoub 2024-03-15
> RegexInterpreter's Goto() and Backtrack() to call EnsureStorage() unconditionally

Thanks. Are you able to measure what impact that has on perf? Hopefully the call would be inlined, and we should end up with effectivlely two comparisons instead of one and it shouldn't be measurable. As long as perf looks good, this general form of fix makes sense.