@rhirano0715 2024-03-16
@stephentoub 
Thank you for checking.

> Are you able to measure what impact that has on perf? 

I'm sorry. I don't understand how to do it.
Is it something that's written around here?
https://github.com/dotnet/runtime/blob/7740dbd7fdb54e87c6aec5eb7b193bc7d21b68f1/docs/workflow/README.md?plain=1#L89-L92