@steveharter 2024-03-27
>I'm sorry. I don't understand how to do it.
Is it something that's written around here?

The common ways to verify perf:
- Add or find an existing RegEx benchmark in https://github.com/dotnet/performance/tree/main/src/benchmarks/micro/libraries/System.Text.RegularExpressions and run before\after.
- Run a temporary benchmark such as by using https://github.com/dotnet/BenchmarkDotNet which is also mentioned in
https://github.com/dotnet/performance/blob/main/docs/microbenchmark-design-guidelines.md